requirements.txt
anthropic>=0.18.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pandas>=2.0.0
flask>=3.0.0
flask-wtf>=1.2.0
flask-limiter>=3.5.0
markupsafe>=2.1.0
gunicorn>=21.2.0
pytest>=7.4.0
